backend/worker/requirements.txt
# Lambda worker dependencies
python-whois==0.9.6
dnspython==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
phonenumbers==8.13.26
psycopg2-binary==2.9.11
sqlalchemy==2.0.25
boto3==1.34.34
httpx==0.26.0
python-dateutil==2.8.2
openai>=1.0.0

